YouTube parameter "pp"
Adds the "pp" query param. This query value is base64 coded and typically holds some control characters as well as the query that the user last searched. Regularly reaches lengths of 30+, which clutters the link, and can also be unique to the user which will allow this parameter to be used as tracking.